Cleaned types
