+++ packages/commonwealth/client/scripts/views/modals/AuthModal/types.ts
@@ -1,18 +1,15 @@
 import { ChainBase } from '@hicommonwealth/shared';
+import { ReactNode } from 'react';
 import { AuthSSOs, AuthWallets } from '../../components/AuthButton/types';
 
-export type AuthModalTabs = {
+export type ModalBaseTabs = {
   name: string;
   options: AuthWallets[] | AuthSSOs[];
 };
 
-export type CreateAccountModal = {
-  onClose: () => void;
-};
-
-export type SignInModalProps = {
+export type ModalVariantProps = {
   onClose: () => any;
-  onSuccess?: () => any;
+  onSuccess?: (isNewAccount?: boolean) => any;
   showWalletsFor?:
     | ChainBase.Ethereum
     | ChainBase.CosmosSDK
@@ -20,7 +17,16 @@ export type SignInModalProps = {
     | ChainBase.Substrate;
 };
 
+export type ModalBaseProps = {
+  layoutType: 'create-account' | 'sign-in';
+  hideDescription?: boolean;
+  customBody?: ReactNode;
+  showAuthenticationOptionsFor?: ('wallets' | 'sso')[];
+  bodyClassName?: string;
+  onSignInClick?: () => void;
+} & ModalVariantProps;
+
 export type AuthModalProps = {
   isOpen: boolean;
   type?: 'create-account' | 'sign-in';
-} & SignInModalProps;
+} & ModalVariantProps;
